refactor: Separate context file into Provider and Consumer and add error handling
