+++ packages/view/src/context/GlobalDataProvider.tsx
@@ -1,0 +1,38 @@
+import type { PropsWithChildren } from "react";
+import { useMemo, useState } from "react";
+
+import { GlobalDataContext, type DateFilterRange } from "hooks";
+import type { ClusterNode } from "types";
+
+export const GlobalDataProvider = ({ children }: PropsWithChildren) => {
+  const [data, setData] = useState<ClusterNode[]>([]);
+  const [filteredData, setFilteredData] = useState<ClusterNode[]>(data);
+  const [selectedData, setSelectedData] = useState<ClusterNode[]>([]);
+  const [filteredRange, setFilteredRange] = useState<DateFilterRange>(undefined);
+  const [loading, setLoading] = useState(false);
+
+  const fetchAnalyzedData = (analyzedData: ClusterNode[]) => {
+    setData(analyzedData);
+    setFilteredData([...analyzedData]);
+    setSelectedData([]);
+    setLoading(false);
+  };
+
+  const value = useMemo(
+    () => ({
+      data,
+      filteredRange,
+      setFilteredRange,
+      filteredData,
+      setFilteredData,
+      selectedData,
+      setSelectedData,
+      fetchAnalyzedData,
+      loading,
+      setLoading,
+    }),
+    [data, filteredRange, filteredData, selectedData, loading]
+  );
+
+  return <GlobalDataContext.Provider value={value}>{children}</GlobalDataContext.Provider>;
+};
