+++ packages/view/src/hooks/useGlobalData.tsx
@@ -1,5 +1,6 @@
-import type { Dispatch, ReactNode } from "react";
-import React, { createContext, useContext, useMemo, useState } from "react";
+import type { Dispatch } from "react";
+import type React from "react";
+import { createContext, useContext } from "react";
 
 import type { ClusterNode } from "types";
 
@@ -23,48 +24,13 @@ type GlobalDataState = {
   setLoading: Dispatch<React.SetStateAction<boolean>>;
 };
 
-export const GlobalDataContext = createContext<GlobalDataState>({} as GlobalDataState);
-
-export const GlobalDataProvider = ({ children }: { children: ReactNode }) => {
-  const [data, setData] = useState<ClusterNode[]>([]);
-  const [filteredData, setFilteredData] = useState<ClusterNode[]>(data);
-  const [selectedData, setSelectedData] = useState<ClusterNode[]>([]);
-  const [filteredRange, setFilteredRange] = useState<DateFilterRange>(undefined);
-  const [loading, setLoading] = useState(false);
-
-  const fetchAnalyzedData = (analyzedData: ClusterNode[]) => {
-    setData(analyzedData);
-    setFilteredData([...analyzedData]);
-    setSelectedData([]);
-    setLoading(false);
-  };
-
-  const value = useMemo(
-    () => ({
-      data,
-      filteredRange,
-      setFilteredRange,
-      filteredData,
-      setFilteredData,
-      selectedData,
-      setSelectedData,
-      fetchAnalyzedData,
-      loading,
-      setLoading,
-    }),
-    [data, filteredRange, filteredData, selectedData, loading]
-  );
-
-  return <GlobalDataContext.Provider value={value}>{children}</GlobalDataContext.Provider>;
-};
+export const GlobalDataContext = createContext<GlobalDataState | undefined>(undefined);
 
 export const useGlobalData = () => {
-  const globalData = useContext<GlobalDataState>(GlobalDataContext);
+  const globalData = useContext(GlobalDataContext);
+  if (!globalData) {
+    throw new Error("Cannot find GlobalDataProvider");
+  }
 
-  return {
-    ...globalData,
-    data: globalData?.data ?? [],
-    filteredData: globalData?.filteredData ?? [],
-    selectedData: globalData?.selectedData ?? null,
-  };
+  return globalData;
 };
